Moved the class MatrixOperatons to a document and modified main to import the Class
